requirements4.txt
django==3.0.7
Pillow==7.2.0
djangorestframework==3.10.3
django-rest-auth==0.9.5
python-dotenv==0.13.0
django-cors-headers==3.4.0
django-allauth==0.42.0
asgiref==3.2.3
psycopg2==2.8.3
psycopg2-binary==2.8.3
gunicorn==19.7.1
django-storages==1.9.1
boto3==1.14.11


